Refactor AgentProxy
Move AgentProxy to a dedicated folder.
Add a few constants.
