+++ include/orwell/game/Game.hpp
@@ -52,12 +52,13 @@ public:
 //	std::shared_ptr< com::Sender > getPublisher();
 
 	std::shared_ptr<Robot> accessRobot(std::string const & iRobotName);
+	std::shared_ptr<Robot const> getRobot(std::string const & iRobotName) const;
 	std::shared_ptr<Robot> accessRobotById(std::string const & iRobotId);
 	bool getHasRobotById(std::string const & iRobotId) const;
 	std::map<std::string, std::shared_ptr<Robot> > const & getRobots() const;
 
 	std::shared_ptr< Player > accessPlayer(std::string const & iPlayerName);
-	std::map< std::string, std::shared_ptr< Player > > const & getPlayers();
+	std::map< std::string, std::shared_ptr< Player > > const & getPlayers() const;
 
 	bool getIsRunning() const;
 
@@ -173,7 +174,8 @@ public:
 
 	std::vector< orwell::game::Landmark > const & getMapLimits() const;
 
-	std::weak_ptr< orwell::game::item::FlagDetector > getFlagDetector(std::string const & iRobotId);
+	std::weak_ptr< orwell::game::item::FlagDetector > getFlagDetector(
+			std::string const & iRobotId);
 private:
 	/// \return
 	///  A RobotID that is not already used.
