+++ include/orwell/proxy/AgentProxy.hpp
@@ -4,12 +4,14 @@
 #include <string>
 #include <stdint.h>
 
-#include "orwell/IAgentProxy.hpp"
+#include "orwell/proxy/IAgentProxy.hpp"
 
 namespace orwell
 {
 class Application;
 
+namespace proxy
+{
 enum class OutputMode;
 
 class SwitchOutputMode;
@@ -29,91 +31,62 @@ public :
 			std::string & ioReply) override;
 
 	/// stop application
-	void stopApplication() override;
-
-	/// List all the teams present.
-	///
-	void listTeam(std::string & ioReply) const override;
+	void stopApplication();
 
 	/// add team <name>
 	void addTeam(
-			std::string const & iTeamName) override;
+			std::string const & iTeamName);
 
 	/// remove team <name>
-	void removeTeam(std::string const & iTeamName) override;
-
-	/// get property <property> of team <name>
-	bool getTeam(
-			std::string const & iTeamName,
-			std::string const & iProperty,
-			std::string & oValue) const override;
+	void removeTeam(std::string const & iTeamName);
 
 	/// set property <property> of team
 	void setTeam(
 			std::string const & iTeamName,
 			std::string const & iProperty,
-			std::string const & iValue) override;
-
-	/// List all the robots present.
-	///
-	void listRobot(std::string & ioReply) const override;
+			std::string const & iValue);
 
 	/// add robot <name>
 	void addRobot(
 			std::string const & iRobotName,
-			std::string const & iTeamName) override;
+			std::string const & iTeamName);
 
 	/// remove robot <name>
-	void removeRobot(std::string const & iRobotName) override;
+	void removeRobot(std::string const & iRobotName);
 
 	/// register robot <name>
-	void registerRobot(std::string const & iRobotName) override;
+	void registerRobot(std::string const & iRobotName);
 
 	/// unregister robot <name>
-	void unregisterRobot(std::string const & iRobotName) override;
+	void unregisterRobot(std::string const & iRobotName);
 
 	/// set property <property> of robot <name>
 	void setRobot(
 			std::string const & iRobotName,
 			std::string const & iProperty,
-			std::string const & iValue) override;
-
-	/// get property <property> of robot <name>
-	bool getRobot(
-			std::string const & iRobotName,
-			std::string const & iProperty,
-			std::string & oValue) const override;
-
-	/// List all the players present.
-	///
-	void listPlayer(std::string & ioReply) const override;
+			std::string const & iValue);
 
 	/// add player <name>
-	void addPlayer(std::string const & iPlayerName) override;
+	void addPlayer(std::string const & iPlayerName);
 
 	/// remove player <name>
-	void removePlayer(std::string const & iPlayerName) override;
+	void removePlayer(std::string const & iPlayerName);
 
 	/// start game
-	void startGame() override;
+	void startGame();
 
 	/// stop game
-	void stopGame() override;
-
-	/// get property <property> of game
-	bool getGame(
-			std::string const & iProperty,
-			std::string & oValue) const override;
+	void stopGame();
 
 	/// set property <property> of game
 	void setGame(
 			std::string const & iProperty,
-			uint32_t const iValue) override;
+			uint32_t const iValue);
 
 	/// view a team in details
 	void viewTeam(
 			std::string const & iName,
-			std::string & oReply) const override;
+			std::string & oReply) const;
 
 protected :
 	friend class SwitchOutputMode;
@@ -127,4 +100,6 @@ private :
 	orwell::Application & m_application;
 	OutputMode m_outputMode;
 };
+
+}
 }
