+++ include/orwell/proxy/OutputMode.hpp
@@ -1,0 +1,19 @@
+#pragma once
+
+#include <iosfwd>
+
+namespace orwell
+{
+namespace proxy
+{
+
+enum class OutputMode
+{
+	kText,
+	kJson,
+};
+
+std::ostream & operator<<(std::ostream & ioStream, OutputMode const iOutputMode);
+
+}
+}
