--- src/orwell/IAgentProxy.cpp
@@ -1,10 +1,0 @@
-#include "orwell/IAgentProxy.hpp"
-
-namespace orwell
-{
-
-IAgentProxy::~IAgentProxy()
-{
-}
-
-}
