+++ src/orwell/proxy/IAgentProxy.cpp
@@ -1,0 +1,13 @@
+#include "orwell/proxy/IAgentProxy.hpp"
+
+namespace orwell
+{
+namespace proxy
+{
+
+IAgentProxy::~IAgentProxy()
+{
+}
+
+}
+}
